Create LocalStatus or LocalStatusSQLite folders
Depending on the configuration we use the plain text or the new
experimental sqlite backend for the LocalStatus cache. Make plain text
the default status backend but allow people to configure
status_backend=sqlite in their local [Repository ...] section.

(It would also make sense to put this option in the [Account ...]
section)

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>